+++ plugin-gradle/CHANGELOG.md
@@ -3,6 +3,9 @@
 We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format.
 
 ## [Unreleased]
+### Added
+- `equoIde` now downloads its dependencies only if it is called directly. This means that CI builds don't need to download IDE dependencies. ([#89]https://github.com/equodev/equo-ide/pull/89))
+  - Also, `equoIde` no longer adds `mavenCentral()` automatically.
 
 ## [0.15.0] - 2023-02-19
 ### Added