+++ plugin-gradle/CHANGELOG.md
@@ -10,6 +10,7 @@ We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format.
 - `equoIde` now downloads its dependencies only if it is called directly. This means that CI builds don't need to download IDE dependencies. ([#89]https://github.com/equodev/equo-ide/pull/89))
   - Also, `equoIde` no longer adds `mavenCentral()` automatically.
 - Gradle DSL now includes `rust`. ([#94](https://github.com/equodev/equo-ide/pull/94))
+- Gradle DSL now includes `m2e`. ([#95](https://github.com/equodev/equo-ide/pull/95))
 
 ## [0.15.0] - 2023-02-19
 ### Added